refactor: split frontend_select files into separate per-screen source files
